ai_powered_tests/requirements.txt
# Browser Use requirements
browser-use>=0.1.14
langchain>=0.3.0
langchain-anthropic>=0.2.5
langchain-openai>=0.2.5
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# For parallel execution
asyncio>=3.4.3
aiohttp>=3.9.0

# Existing Playwright for comparison
playwright>=1.48.0